Deduplicate some debugger/decoder logic
