+++ packages/debugger/lib/evm/actions/index.js
@@ -3,6 +3,7 @@ export function addContext({
   contractName,
   binary,
   sourceMap,
+  primarySource,
   compiler,
   compilationId,
   abi,
@@ -15,6 +16,7 @@ export function addContext({
     contractName,
     binary,
     sourceMap,
+    primarySource,
     compiler,
     compilationId,
     abi,
